Improve abilities
